Add OpenCV integration attempt
Okay, so I think everything is working here that needs to, rather it is
just my specific installtion of OpenCV that is out of whack. I'll get
back to this tomorrow morning. For now I'll share this to see if you
guys get any luck installing OpenCV and running this.