Add a lot...
Not surprisingly, it's taking me a little longer than I thought to get
everything up and running. I believe I have written everything except
for the algorithm which attempts a few thousand guesses  for what the
pose is. (These guesses are based on Jiang's calculation of the two
possible locations for depth of a bone's endpoint given an orthographic
projection)